+++ components/raftstore-v2/src/operation/pd.rs
@@ -2,20 +2,17 @@
 
 //! This module implements the interactions with pd.
 
-use std::{sync::atomic::Ordering, time::Instant};
+use std::sync::atomic::Ordering;
 
 use engine_traits::{KvEngine, RaftEngine};
 use fail::fail_point;
 use kvproto::{metapb, pdpb};
 use raftstore::store::{
-    metrics::{RAFT_PEER_PENDING_DURATION, STORE_SNAPSHOT_TRAFFIC_GAUGE_VEC},
+    metrics::STORE_SNAPSHOT_TRAFFIC_GAUGE_VEC,
     Transport,
 };
 use slog::{debug, error};
-use tikv_util::{
-    slog_panic,
-    time::{duration_to_sec, InstantExt},
-};
+use tikv_util::{slog_panic, time::Instant};
 
 use crate::{
     batch::StoreContext,
@@ -247,70 +244,5 @@ impl<EK: KvEngine, ER: RaftEngine> Peer<EK, ER> {
                 "err" => %e,
             );
         }
-    }
-}
-
-#[derive(Default)]
-pub struct AbnormalPeerContext {
-    /// Record the instants of peers being added into the configuration.
-    /// Remove them after they are not pending any more.
-    peers_start_pending_time: Vec<(u64, Instant)>,
-    /// A inaccurate cache about which peer is marked as down.
-    down_peers: Vec<u64>,
-}
-
-impl AbnormalPeerContext {
-    #[inline]
-    pub fn is_empty(&self) -> bool {
-        self.peers_start_pending_time.is_empty() && self.down_peers.is_empty()
-    }
-
-    #[inline]
-    pub fn reset(&mut self) {
-        self.peers_start_pending_time.clear();
-        self.down_peers.clear();
-    }
-
-    #[inline]
-    pub fn is_peer_down(&self, peer_id: u64) -> bool {
-        if self.down_peers.is_empty() {
-            return false;
-        }
-        self.down_peers.contains(&peer_id)
-    }
-
-    #[inline]
-    pub fn is_peer_pending(&self, peer_id: u64) -> bool {
-        if self.peers_start_pending_time.is_empty() {
-            return false;
-        }
-        self.peers_start_pending_time.iter().any(|p| p.0 == peer_id)
-    }
-
-    #[inline]
-    pub fn collect_down_peers(&mut self, peer_ids: Vec<u64>) {
-        self.down_peers = peer_ids;
-    }
-
-    #[inline]
-    pub fn append_pending_peers(&mut self, mut peers: Vec<(u64, Instant)>) {
-        self.peers_start_pending_time.append(&mut peers);
-    }
-
-    #[inline]
-    pub fn retain_pending_peers(&mut self, f: impl FnMut(&mut (u64, Instant)) -> bool) -> bool {
-        let len = self.peers_start_pending_time.len();
-        self.peers_start_pending_time.retain_mut(f);
-        len != self.peers_start_pending_time.len()
-    }
-
-    pub fn observe_pending_peers(&self) {
-        let _ = self
-            .peers_start_pending_time
-            .iter()
-            .map(|(_, pending_after)| {
-                let elapsed = duration_to_sec(pending_after.saturating_elapsed());
-                RAFT_PEER_PENDING_DURATION.observe(elapsed);
-            });
     }
 }
