+++ components/raftstore-v2/src/operation/ready/mod.rs
@@ -300,9 +300,8 @@ impl<EK: KvEngine, ER: RaftEngine> Peer<EK, ER> {
         }
 
         // There are two different cases to check peers can be bring back.
-        // 1. If the peer is pending, then only AppendResponse can bring it back to not
-        // pending. 2. If the peer is down, then HeartbeatResponse and
-        // AppendResponse can bring it back to up.
+        // 1. If the peer is pending, then only AppendResponse can bring it back to up.
+        // 2. If the peer is down, then HeartbeatResponse and AppendResponse can bring it back to up.
         if self.any_new_peer_catch_up(from_peer_id) {
             self.region_heartbeat_pd(ctx)
         }
